+++ crates/chaincash_offchain/src/contracts.rs
@@ -2,17 +2,17 @@ use ergo_lib::ergoscript_compiler::compiler::compile;
 use ergo_lib::ergotree_ir::ergo_tree::ErgoTree;
 use once_cell::sync::Lazy;
 
-pub(crate) static RESERVE_CONTRACT: &str =
+pub static RESERVE_CONTRACT: &str =
     include_str!("../../../contracts/chaincash/contracts/onchain/reserve.es");
 // Not currently able to compile with sigma-rust, using node api instead
-pub(crate) static RESERVE_ERGO_TREE: Lazy<ErgoTree> =
+pub static RESERVE_ERGO_TREE: Lazy<ErgoTree> =
     Lazy::new(|| compile(RESERVE_CONTRACT, Default::default()).unwrap());
 
-pub(crate) static RECEIPT_CONTRACT: &str =
+pub static RECEIPT_CONTRACT: &str =
     include_str!("../../../contracts/chaincash/contracts/onchain/receipt.es");
 
-pub(crate) static NOTE_CONTRACT: &str =
+pub static NOTE_CONTRACT: &str =
     include_str!("../../../contracts/chaincash/contracts/onchain/note.es");
 // Not currently able to compile with sigma-rust, using node api instead
-pub(crate) static NOTE_ERGO_TREE: Lazy<ErgoTree> =
+pub static NOTE_ERGO_TREE: Lazy<ErgoTree> =
     Lazy::new(|| compile(NOTE_CONTRACT, Default::default()).unwrap());
