+++ crates/chaincash_offchain/src/lib.rs
@@ -1,3 +1,4 @@
+pub mod boxes;
 pub mod contracts;
 pub mod node;
 pub mod transactions;
@@ -5,7 +6,6 @@ pub mod transactions;
 use thiserror::Error;
 
 pub use ergo_client::Error as ClientError;
-pub use transactions::TransactionService;
 
 #[derive(Debug, Error)]
 pub enum Error {
@@ -15,6 +15,6 @@ pub enum Error {
     #[error("Ergo Node client error")]
     Node(#[from] ergo_client::node::NodeError),
 
-    #[error("transaction error: {0}")]
+    #[error("Transaction error occurred")]
     Transaction(#[from] transactions::TransactionError),
 }
