+++ crates/chaincash_offchain/src/transactions.rs
@@ -53,102 +53,3 @@ pub struct TxContext {
     change_address: String,
     fee: u64,
 }
-
-#[derive(Clone)]
-pub struct TransactionService<'a> {
-    node: &'a NodeClient,
-}
-
-impl<'a> TransactionService<'a> {
-    pub fn new(node: &'a NodeClient) -> Self {
-        Self { node }
-    }
-
-    async fn box_selection_with_amount(
-        &self,
-        amount: u64,
-    ) -> Result<BoxSelection<ErgoBox>, crate::Error> {
-        let inputs = self
-            .node
-            .extensions()
-            .get_utxos_summing_amount(amount)
-            .await?;
-        // kinda irrelevant since we already have suitable boxes but box selectors required by ergo-lib txbuilder
-        Ok(SimpleBoxSelector::new()
-            .select(
-                inputs,
-                amount.try_into().map_err(TransactionError::from)?,
-                &[],
-            )
-            .map_err(TransactionError::from)?)
-    }
-
-    async fn get_tx_ctx(&self) -> Result<TxContext, crate::Error> {
-        let wallet_status = self.node.endpoints().wallet()?.status().await?;
-        let info = self.node.endpoints().root()?.info().await?;
-
-        if wallet_status.change_address.is_empty() {
-            Err(TransactionError::ChangeAddress(
-                "address not set".to_owned(),
-            ))?
-        } else {
-            Ok(TxContext {
-                current_height: info.full_height as u32,
-                change_address: wallet_status.change_address,
-                fee: *SUGGESTED_TX_FEE().as_u64(),
-            })
-        }
-    }
-
-    pub async fn mint_reserve(&self, request: MintReserveRequest) -> Result<String, crate::Error> {
-        let ctx = self.get_tx_ctx().await?;
-        let selected_inputs = self
-            .box_selection_with_amount(request.amount + ctx.fee)
-            .await?;
-        let reserve_tree = self
-            .node
-            .extensions()
-            .compile_contract(RESERVE_CONTRACT)
-            .await?;
-        let unsigned_tx = mint_reserve_transaction(request, reserve_tree, selected_inputs, ctx)?;
-        let submitted_tx = self.node.extensions().sign_and_submit(unsigned_tx).await?;
-        // todo, add reserve to db
-        // should return minted reserve?
-        Ok(submitted_tx.id().to_string())
-    }
-
-    pub async fn mint_note(&self, request: MintNoteRequest) -> Result<String, crate::Error> {
-        let reserve_tree_bytes = self
-            .node
-            .extensions()
-            .compile_contract(RESERVE_CONTRACT)
-            .await?
-            .sigma_serialize_bytes()
-            .unwrap();
-        let reserve_hash = bs58::encode(blake2b256_hash(&reserve_tree_bytes[1..])).into_string();
-        let receipt_contract = RECEIPT_CONTRACT.replace("$reserveContractHash", &reserve_hash);
-        let receipt_tree_bytes = self
-            .node
-            .extensions()
-            .compile_contract(&receipt_contract)
-            .await?
-            .sigma_serialize_bytes()
-            .unwrap();
-        let receipt_hash = bs58::encode(blake2b256_hash(&receipt_tree_bytes[1..])).into_string();
-        let ctx = self.get_tx_ctx().await?;
-        let selected_inputs = self
-            .box_selection_with_amount(BoxValue::SAFE_USER_MIN.as_u64() + ctx.fee)
-            .await?;
-        let note_contract = NOTE_CONTRACT
-            .replace("$reserveContractHash", &reserve_hash)
-            .replace("$receiptContractHash", &receipt_hash);
-        let contract_tree = self
-            .node
-            .extensions()
-            .compile_contract(&note_contract)
-            .await?;
-        let unsigned_tx = mint_note_transaction(request, contract_tree, selected_inputs, ctx)?;
-        let submitted_tx = self.node.extensions().sign_and_submit(unsigned_tx).await?;
-        Ok(submitted_tx.id().to_string())
-    }
-}
