+++ crates/chaincash_services/Cargo.toml
@@ -1,16 +1,13 @@
 [package]
-name = "chaincash_offchain"
+name = "chaincash_services"
 version.workspace = true
 edition.workspace = true
 license.workspace = true
 
 [dependencies]
-once_cell = "1.18.0"
+bs58 = "0.5.0"
 ergo_lib = { workspace = true }
 ergo_client = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
+chaincash_store = { path = "../chaincash_store" }
+chaincash_offchain = { path = "../chaincash_offchain" }
 thiserror = { workspace = true }
-ergo_avltree_rust = "0.1.0"
-sigma-ser = "0.13.0"
-tracing = { workspace = true }
