--- crates/chaincash_specs/Cargo.toml
@@ -1,9 +1,0 @@
-[package]
-name = "chaincash_specs"
-version.workspace = true
-edition.workspace = true
-license.workspace = true
-
-[dependencies]
-ergo_lib = { workspace = true }
-thiserror = { workspace = true }
