+++ crates/chaincash_store/Cargo.toml
@@ -5,7 +5,7 @@ edition.workspace = true
 license.workspace = true
 
 [dependencies]
-chaincash_specs = { path = "../chaincash_specs" }
+chaincash_offchain = { path = "../chaincash_offchain" }
 diesel = { version = "2.1.3", features = ["sqlite", "r2d2", "returning_clauses_for_sqlite_3_35"] }
 diesel_migrations = "2.1.0"
 thiserror = { workspace = true }
